Make use of the Supervisor API on device shutdown and/or reboot.

+++ server.py
@@ -25,6 +25,7 @@ from functools import wraps
 from hurry.filesize import size
 from mimetypes import guess_type, guess_extension
 from os import getenv, listdir, makedirs, mkdir, path, remove, rename, statvfs, stat, walk
+from retry.api import retry_call
 from subprocess import check_output
 from urlparse import urlparse
 
@@ -45,16 +46,20 @@ from lib import raspberry_pi_helper
 
 from lib.github import is_up_to_date
 from lib.auth import authorized
-from lib.utils import download_video_from_youtube, json_dump
-from lib.utils import generate_perfect_paper_password, is_docker
-from lib.utils import get_active_connections, remove_connection
-from lib.utils import get_node_ip, get_node_mac_address
-from lib.utils import get_video_duration
-from lib.utils import is_balena_app, is_demo_node
-from lib.utils import string_to_bool
-from lib.utils import connect_to_redis
-from lib.utils import url_fails
-from lib.utils import validate_url
+
+from lib.utils import (
+    download_video_from_youtube, json_dump,
+    generate_perfect_paper_password, is_docker,
+    get_active_connections, remove_connection,
+    get_node_ip, get_node_mac_address,
+    get_video_duration,
+    is_balena_app, is_demo_node,
+    shutdown_via_balena_supervisor, reboot_via_balena_supervisor,
+    string_to_bool,
+    connect_to_redis,
+    url_fails,
+    validate_url,
+)
 
 from settings import CONFIGURABLE_SETTINGS, DEFAULTS, LISTEN, PORT, settings, ZmqPublisher, ZmqCollector
 
@@ -106,7 +111,10 @@ def reboot_screenly():
     """
     Background task to reboot Screenly-OSE.
     """
-    r.publish('hostcmd', 'reboot')
+    if is_balena_app():
+        retry_call(reboot_via_balena_supervisor, tries=5, delay=1)
+    else:
+        r.publish('hostcmd', 'reboot')
 
 
 @celery.task
@@ -114,7 +122,10 @@ def shutdown_screenly():
     """
     Background task to shutdown Screenly-OSE.
     """
-    r.publish('hostcmd', 'shutdown')
+    if is_balena_app():
+        retry_call(shutdown_via_balena_supervisor, tries=5, delay=1)
+    else:
+        r.publish('hostcmd', 'shutdown')
 
 
 @celery.task
